Nicer mountains.

+++ WorldJourney/InfiniteNoise.h
@@ -5,6 +5,6 @@ using namespace metal;
 #ifndef InfiniteNoise_h
 #define InfiniteNoise_h
 
-float4 fbmInf3(int3 cubeOrigin, int cubeSize, float3 x, float f, float a, int o, float sharpness);
+float4 fbmInf3(int3 cubeOrigin, int cubeSize, float3 x, float f, float a, float o, float sharpness);
 
 #endif
