+++ WorldJourney/WorldTerrain.h
@@ -1,6 +1,6 @@
 #ifndef WorldTerrain_h
 #define WorldTerrain_h
 
-float4 sampleInf(int3 cubeOrigin, int cubeSize, float3 x, float a, int o, float t);
+float4 sampleInf(int3 cubeOrigin, int cubeSize, float3 x, float a, float o, float t);
 
 #endif
